fix LL to not set prior
LL set prior to false for ML, which was seen as an actual prior, and a warning was raised (as it should). The warning was intercepted by estimate, which consequently fell back on an EAP estimate.

Result; ML estimates were impossible. Fixed by not letting LL set prior at all, and let prob figure it out from context.